requirements.txt
llama-cpp-python==0.2.79
sentence-transformers==2.7.0
transformers==4.42.3
torch==2.3.1
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.5.1
Pillow==10.4.0
requests==2.32.3
beautifulsoup4==4.12.3
python-dotenv==1.0.1
fastapi==0.111.0
uvicorn==0.30.1
textstat==0.7.11
nltk==3.8.1